api_service/requirements.txt
fastapi
uvicorn
influxdb-client
python-jose[cryptography]
python-multipart
pytest
pytest-cov
